pos_receipt_replace_user_by_trigram: Drop compute function
The compute function does not allow any customization on how the
trigram had to be generated, and there a multiple ways to define
how trigram must be generated.

Moreover, the compute did generate duplicates and there was no way
to fix manually since the field was not stored. Also, making it
stored but still computed did risk unwanted updates in case name
was updated.

Having a basic stored field allows each company to customize how
trigram can be generated using a server action. The logic previously
provided is then moved to a cron that can be customized to suit
trigram generation for each company.